requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Vector database
chromadb==0.4.22
numpy<2.0.0  # ChromaDB is not compatible with NumPy 2.0+

# PDF/DOCX text extraction
pypdf2==3.0.1
python-docx==1.1.0

# LLM and transformers (CUDA 12.8 compatible)
# Note: Install PyTorch nightly with CUDA 12.8 support first:
# pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128
torch>=2.4.0
transformers>=4.40.0  # Required for Llama-3.2 compatibility
accelerate>=0.25.0  # Required for device_map support
bitsandbytes==0.41.3
sentencepiece==0.1.99

# Text processing
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0


